refactor: remove code flow tenant login
